Clean build directly only when explicitly called for.

+++ build.sh
@@ -1,6 +1,10 @@
 #!/bin/sh
 PROJECT_DIR=`dirname $0`
-rm -rf $PROJECT_DIR/build
-mkdir $PROJECT_DIR/build
+
+if [ "$1" == "--clean" ]; then
+  rm -rf $PROJECT_DIR/build
+  mkdir $PROJECT_DIR/build
+fi
+
 cmake -H$PROJECT_DIR -B$PROJECT_DIR/build
 make -C $PROJECT_DIR/build
